+++ product-service/src/main/java/com/example/shopping/productservice/app/product/web/ProductController.java
@@ -3,6 +3,8 @@ package com.example.shopping.productservice.app.product.web;
 import com.example.shopping.productservice.app.product.domain.Product;
 import com.example.shopping.productservice.app.product.dto.ProductRequest;
 import com.example.shopping.productservice.app.product.service.ProductService;
+import com.example.shopping.productservice.exception.ErrorUtil;
+import com.example.shopping.productservice.exception.ProductServiceException;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -43,8 +45,16 @@ public class ProductController {
 //        return ResponseEntity.status(HttpStatus.OK).body(productService.orderProduct(product));
 //    }
 
+//    @GetMapping(value = "/product/{productId}")
+//    public ResponseEntity<Product> getProductById(@PathVariable("productId") Long productId) {
+//        Product product = productService.findById(productId);
+//        return ResponseEntity.status(HttpStatus.OK).body(product);
+//    }
     @GetMapping(value = "/product/{productId}")
     public ResponseEntity<?> getProductById(@PathVariable("productId") Long productId) {
+        if (productId == null) {
+            throw new ProductServiceException("No product Id");
+        }
         Product product = productService.findById(productId);
         return ResponseEntity.status(HttpStatus.OK).body(product);
     }
@@ -56,8 +66,16 @@ public class ProductController {
         return ResponseEntity.status(HttpStatus.OK).body(productService.findById(productId));
     }
 
+//    @DeleteMapping(value = "/product/{productId}/delete")
+//    public ResponseEntity<?> deleteProduct(@PathVariable("productId") Long productId) {
+//        productService.deleteProduct(productId);
+//        return new ResponseEntity<>(HttpStatus.OK);
+//    }
     @DeleteMapping(value = "/product/{productId}/delete")
     public ResponseEntity<?> deleteProduct(@PathVariable("productId") Long productId) {
+        if (productId == null) {
+            return ErrorUtil.create(HttpStatus.BAD_REQUEST,"bad request", "no path variable");
+        }
         productService.deleteProduct(productId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -66,11 +84,6 @@ public class ProductController {
     public Page<Product> getProductByPagination(@RequestParam(name = "page", defaultValue = "0") Integer page,
                                                 @RequestParam(name = "size", defaultValue = "10") Integer size) {
         return productService.findAllByPage(page, size);
-    }
-
-    @GetMapping
-    public ResponseEntity<List<Product>> getAllProducts() {
-        return ResponseEntity.status(HttpStatus.OK).body(productService.findAll());
     }
 
     // Find all by multiple product Ids
